refactor hooks to components
